Changed hierarchy to PlayerShip & Bullet inheriting from ColoredObj Got firing working Got Bullet colors working
