+++ Digital 1/Assets/Prefabs/Player.prefab
@@ -15,7 +15,7 @@ GameObject:
   - component: {fileID: 3606396754197552547}
   m_Layer: 0
   m_Name: Player
-  m_TagString: Untagged
+  m_TagString: Player
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
