+++ Digital 1/Assets/Scripts/ColoredObj.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ae286f79a6208c4dbf8950d51891a3a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
